Fix PatchExitCommand in dynamic mode
It was generating completely wrong data for the Kamek files. Properly fixing it required making Command.Address nullable, which proceeded to affect all the rest of the Command subclasses. I've added null checks where appropriate -- notably, not in subclasses which don't allow Address to be initialized from a nullable constructor parameter.