Fix loader v2 on hardware
We need to ensure that, at minimum, the instruction cache is completely invalidated for the entire region we copy to before we jump to it. Otherwise the processor will see stale instruction data and crash when we jump to it.

This bug existed before this PR, technically, but the PR triggers it consistently because the loader immediately jumps to ctors instead of jumping back to the game code and just letting hooks execute sometime later.

I may have also added more cache syncing/invalidation than strictly necessary, because I don't fully understand exactly what each of the instructions does even after reading about them in 6xx_pem and I want to err on the side of safety. The performance hit will most likely be small for most patches.