NSMBW Loader: add support for Korean and Taiwanese versions
The new address maps for these versions are accurate down to the opcode level (i.e. if you map any code address across, the opcode will match -- albeit with a caveat explained in a comment at the top of the file).

For version detection, an additional test has to be added because K and W are the only two versions that can't be distinguished with the address used for the main test. I'm only testing the third byte instead of the full u32, because the other 3 bytes are the same between the two, and because the smaller immediate value will probably save an instruction or two.

K and W also only have one version each. I refer to it here as version "0", and special-case that version number so it's left out of the engine filenames. "0" isn't entirely arbitrary, by the way -- it matches the "version" field in these version's disc headers (byte 0x07 -- see https://wiibrew.org/wiki/Wii_disc), whereas the traditional versions 1 and 2 really do have "1" and "2" there instead, as opposed to, say, 0 and 1.