Fix hooks in assembly files (#21)
CodeWarrior's assembler deletes local labels that have no references to them, so hooks were being deleted. The ideal solution would be to tell CodeWarrior to preserve the symbols as exports with local binding, but there seems to be no way to do that. There are three possible solutions:

1) declare hooks as global symbols, even though they're not actually globally unique, and modify Kamek to accept these object files even though they're technically wrong

2) create dummy references to the symbols to prevent CodeWarrior from deleting them (kind of ugly, but produces object files that are technically correct, and therefore requires no changes to Kamek itself)

3) modify Kamek to read the .kamek section directly, without using symbols at all

I believe #2 is the best, since this is a CodeWarrior issue rather than a Kamek one, and it accordingly limits the "ugliness" to the compilation phase rather than the linking phase (wherein Kamek itself is doing nothing wrong).

This commit implements that by putting dummy references in a ".discard" section, which Kamek doesn't recognize and hence discards.