Trim published binaries (#25)
(and also fix a formatting mistake in the .csproj file)

Suggested by CLF78. This one-line project config change reduces the artifact sizes for each platform from ~61 MB to ~12 MB, which is a pretty huge improvement. The trimmed Linux binary seems to work fine, and I'd be surprised if this breaks either of the other two.